refactor(protocol-designer): remove hooks.ts for custom hooks
we should not use hooks.ts because that isn't not the best practice for managing custom hooks. hooks
is fine when we work on a prototype or something experimental.
