+++ protocol-designer/src/pages/Designer/ProtocolSteps/StepForm/StepTools/MoveLiquidTools/LiquidClassesStepTools.tsx
@@ -11,7 +11,7 @@ import {
 import type { ChangeEvent, Dispatch, SetStateAction } from 'react'
 import type { FormData } from '../../../../../../form-types'
 import type { FieldPropsByName } from '../../types'
-import type { LiquidClassOption } from './hooks'
+import type { LiquidClassOption } from './hooks/useAssignLiquidClass'
 
 interface LiquidClassesStepToolsProps {
   propsForFields: FieldPropsByName
